feat(wave-8): dynamic horizontal plugin composition for production MCP boot
Move sync compose to compose-horizontal-plugin-layers-static.ts with Static
suffix exports. Barrel re-exports dynamic + static; deprecated aliases keep
tests on sync path.

Add ensureClawqlApi() for dynamic import composition; getClawqlApi() still
sync-bootstraps via static compose. Production server.ts and server-http.ts
await dynamic bootstrap before MCP server creation.
