+++ src/compose-horizontal-plugin-layers.ts
@@ -1,103 +1,36 @@
 /**
- * Dynamic horizontal tier Layer composition for MCP transport and Operator reconciliation (#255).
+ * Horizontal plugin Layer composition — dynamic (production) and static (tests / sync bootstrap).
+ *
+ * @see compose-horizontal-plugin-layers-dynamic.ts
+ * @see compose-horizontal-plugin-layers-static.ts
  */
-import {
-  basePluginCompositionFlags,
-  type ClawqlOptionalToolFlags,
-  type ClawQLApiRuntimeError,
-  type ClawQLApiRuntimeServices,
-} from "clawql-api";
-import { makeAutomationLayer } from "clawql-automation/plugin";
-import { natsConfiguredForConsumer } from "clawql-automation/nats/env";
-import { makeDocumentsLayer } from "clawql-documents/plugin";
-import { makeMemoryLayer } from "clawql-memory/plugin";
-import { makeOntologyLayer } from "clawql-ontology/plugin";
-import { createOuroborosHarnessPlugin, makeHarnessLayer } from "clawql-harness/plugin";
-import { makeSandboxLayer } from "clawql-sandbox/plugin";
-import { makeDataLayer } from "clawql-data/plugin";
-import { makeWebLayer } from "clawql-web/plugin";
-import {
+
+export {
+  composeHorizontalPluginLayersDynamic,
+  composeHorizontalPluginLayersDynamicEffect,
+  composeHorizontalPluginLayersDynamicFromTierSpec,
+  type ComposeHorizontalPluginLayersOptions,
+} from "./compose-horizontal-plugin-layers-dynamic.js";
+
+export {
+  composeHorizontalPluginLayersStatic,
+  composeHorizontalPluginLayersFromTierSpecStatic,
   optionalFlagsFromHorizontalTierSpec,
   type ClawQLHorizontalTierSpec,
-} from "clawql-operator/spec";
-import type { Layer } from "effect";
+} from "./compose-horizontal-plugin-layers-static.js";
 
-export type { ClawQLHorizontalTierSpec } from "clawql-operator/spec";
-export { optionalFlagsFromHorizontalTierSpec } from "clawql-operator/spec";
-
-export type ComposeHorizontalPluginLayersOptions = {
-  /** When true, include NATS workflow worker in AutomationLayer (env-gated by default). */
-  readonly includeNatsWorker?: boolean;
-};
+import {
+  composeHorizontalPluginLayersStatic,
+  composeHorizontalPluginLayersFromTierSpecStatic,
+} from "./compose-horizontal-plugin-layers-static.js";
 
 /**
- * Builds Effect Layers for all enabled horizontal tiers from parsed optional flags.
+ * @deprecated Prefer {@link composeHorizontalPluginLayersDynamic} or {@link ensureClawqlApi}.
  */
-export function composeHorizontalPluginLayers(
-  flags: ClawqlOptionalToolFlags,
-  options: ComposeHorizontalPluginLayersOptions = {}
-): readonly Layer.Layer<never, ClawQLApiRuntimeError, ClawQLApiRuntimeServices>[] {
-  const includeNatsWorker = options.includeNatsWorker ?? natsConfiguredForConsumer();
-  const layers: Layer.Layer<never, ClawQLApiRuntimeError, ClawQLApiRuntimeServices>[] = [];
-  if (flags.enableMemory) {
-    layers.push(makeMemoryLayer());
-  }
-  if (flags.enableDocuments) {
-    layers.push(
-      makeDocumentsLayer({
-        enableOnyx: flags.enableOnyxKnowledge,
-        enableIdpPipeline: flags.enableIdpPipeline,
-        enableIdpClassifier: flags.enableIdpClassifier,
-        enableLangextract: flags.enableLangextract,
-        enablePdfInspector: flags.enablePdfInspector,
-        enableAnydoc: flags.enableAnydoc,
-      })
-    );
-  }
-  if (
-    flags.enableSchedule ||
-    flags.enableNotify ||
-    flags.enableWorkflow ||
-    flags.enableArgoCd ||
-    flags.enableHitlLabelStudio ||
-    includeNatsWorker
-  ) {
-    layers.push(
-      makeAutomationLayer({
-        enableSchedule: flags.enableSchedule,
-        enableNotify: flags.enableNotify,
-        enableWorkflow: flags.enableWorkflow,
-        enableArgoCd: flags.enableArgoCd,
-        enableHitlLabelStudio: flags.enableHitlLabelStudio,
-        enableNatsWorker: includeNatsWorker,
-      })
-    );
-  }
-  if (flags.enableSandbox) {
-    layers.push(makeSandboxLayer());
-  }
-  if (flags.enableData) {
-    layers.push(makeDataLayer());
-  }
-  if (flags.enableWeb) {
-    layers.push(makeWebLayer());
-  }
-  if (flags.enableOntology) {
-    layers.push(makeOntologyLayer({ enableWrites: flags.enableOntologyWrites }));
-  }
-  // Ouroboros is always a clawql-harness plugin (no env / tier enable gate).
-  layers.push(
-    makeHarnessLayer({
-      plugins: [createOuroborosHarnessPlugin({ enableLangfuseEval: flags.enableLangfuseEval })],
-    })
-  );
-  return layers;
-}
+export const composeHorizontalPluginLayers = composeHorizontalPluginLayersStatic;
 
-/** Composes horizontal plugin Layers from a CRD-style tier spec (config — not CLAWQL_ENABLE_*). */
-export function composeHorizontalPluginLayersFromTierSpec(
-  spec: ClawQLHorizontalTierSpec
-): readonly Layer.Layer<never, ClawQLApiRuntimeError, ClawQLApiRuntimeServices>[] {
-  const flags = optionalFlagsFromHorizontalTierSpec(spec, basePluginCompositionFlags());
-  return composeHorizontalPluginLayers(flags);
-}
+/**
+ * @deprecated Prefer {@link composeHorizontalPluginLayersDynamicFromTierSpec} or {@link ensureClawqlApi}.
+ */
+export const composeHorizontalPluginLayersFromTierSpec =
+  composeHorizontalPluginLayersFromTierSpecStatic;
