+++ src/server.ts
@@ -18,7 +18,7 @@ import "./load-env.js";
 import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
 import { NPM_PACKAGE_VERSION } from "./npm-version.js";
 import { loadSpec, registerSpecCacheShutdownHooks } from "clawql-api";
-import { createRegisteredMcpServer } from "./mcp-server-factory.js";
+import { createRegisteredMcpServerAsync } from "./mcp-server-factory.js";
 import { preloadSchemaFieldCacheFromDisk } from "./tools.js";
 import { validateOrDegradeObsidianVaultAtStartup } from "./vault-config.js";
 import { registerPostgresPoolShutdownHooks } from "clawql-memory/vector/pgvector";
@@ -49,7 +49,7 @@ async function main() {
   await ensureProcessWormHostBooted();
   await validateOrDegradeObsidianVaultAtStartup();
 
-  const server = createRegisteredMcpServer({
+  const server = await createRegisteredMcpServerAsync({
     name: "cloudrun-mcp",
     version: NPM_PACKAGE_VERSION,
   });
